+++ Source/EventFlow/Aggregates/IMetadata.cs
@@ -34,6 +34,9 @@ namespace EventFlow.Aggregates
         int AggregateSequenceNumber { get; }
         string AggregateId { get; }
 
+        IMetadata CloneWith(params KeyValuePair<string, string>[] keyValuePairs);
         IMetadata CloneWith(IEnumerable<KeyValuePair<string, string>> keyValuePairs);
+        string GetMetadataValue(string key);
+        T GetMetadataValue<T>(string key, Func<string, T> converter);
     }
 }