+++ Source/EventFlow/Exceptions/MetadataKeyNotFoundException.cs
@@ -21,28 +21,11 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 using System;
-using JetBrains.Annotations;
 
 namespace EventFlow.Exceptions
 {
-    public class DomainError : Exception
+    public class MetadataKeyNotFoundException : ArgumentOutOfRangeException
     {
-        protected DomainError(string message)
-            : base(message) { }
-
-        protected DomainError(string message, Exception innerException)
-            : base(message, innerException) { }
-
-        [StringFormatMethod("format")]
-        public static DomainError With(string format, params object[] args)
-        {
-            return new DomainError(string.Format(format, args));
-        }
-
-        [StringFormatMethod("format")]
-        public static DomainError With(Exception innerException, string format, params object[] args)
-        {
-            return new DomainError(string.Format(format, args), innerException);
-        }
+        public MetadataKeyNotFoundException(string key) : base($"Could not find metadata key '{key}'") { }
     }
 }